// src/models/color.ts
export type ColorKey =
  | 'main-green'
  | 'main-disable'
  | 'main-black'
  | 'main-white'
  | 'main-pink'
  | 'error-50'
  | 'error-100'
  | 'error-200'
  | 'error-300'
  | 'error-400'
  | 'error-500'
  | 'error-600'
  | 'error-700'
  | 'error-800'
  | 'surface-50'
  | 'surface-100'
  | 'surface-200'
  | 'surface-300'
  | 'surface-400'
  | 'surface-500'
  | 'surface-600'
  | 'surface-700'
  | 'surface-800'
  | 'neutral-000'
  | 'neutral-100'
  | 'neutral-200'
  | 'neutral-300'
  | 'neutral-400'
  | 'neutral-500'
  | 'neutral-600'
  | 'neutral-700'
  | 'neutral-800'
  | 'orange-50'
  | 'orange-100'
  | 'orange-200'
  | 'orange-300'
  | 'orange-400'
  | 'orange-500'
  | 'orange-600'
  | 'orange-700'
  | 'orange-800'
  | 'purple-50'
  | 'purple-100'
  | 'purple-200'
  | 'purple-300'
  | 'purple-400'
  | 'purple-500'
  | 'purple-600'
  | 'purple-700'
  | 'purple-800'
  | 'yellow-100'
  | 'profile-coral'
  | 'profile-dark-blue'
  | 'profile-sky-blue'
  | 'profile-violet'
  | 'profile-green'
  | 'blue-100'
  | 'blue-200'
  | 'blue-300';